+++ evolve/evaluation_network_simulator.go
@@ -9,7 +9,7 @@ import (
 )
 
 // perByteEvaluation for evolve with network sim, 1 i32 input, 1 i32 output
-func perByteEvaluation_NetworkSim(ind *cxast.CXProgram, solPrototype *cxast.CXFunction, numberOfRounds int) int64 {
+func perByteEvaluation_NetworkSim(ind *cxast.CXProgram, solPrototype *cxast.CXFunction, numberOfRounds int) float64 {
 	var score int = 0
 	for rounds := 0; rounds < numberOfRounds; rounds++ {
 		// Generate random Input
@@ -28,7 +28,7 @@ func perByteEvaluation_NetworkSim(ind *cxast.CXProgram, solPrototype *cxast.CXFu
 		score = score + countDifferentBits(input, receiverOutput)
 	}
 
-	return int64(score)
+	return float64(score)
 }
 
 // perByteEvaluation for evolve with network sim transmitter, 1 i32 input, 1 i32 output
