+++ evolve/evaluation_odds.go
@@ -10,7 +10,7 @@ import (
 )
 
 // perByteEvaluation for evolve with odd numbers, 1 i32 input, 1 i32 output
-func perByteEvaluation_Odds(ind *cxast.CXProgram, solPrototype *cxast.CXFunction, numberOfRounds int) int64 {
+func perByteEvaluation_Odds(ind *cxast.CXProgram, solPrototype *cxast.CXFunction, numberOfRounds int) float64 {
 	var points int64 = 0
 	var tmp *cxast.CXProgram = cxast.PROGRAM
 	cxast.PROGRAM = ind
@@ -54,5 +54,5 @@ func perByteEvaluation_Odds(ind *cxast.CXProgram, solPrototype *cxast.CXFunction
 	}
 
 	cxast.PROGRAM = tmp
-	return points
+	return float64(points)
 }
