+++ evolve/evaluation_primes.go
@@ -11,7 +11,7 @@ import (
 )
 
 // perByteEvaluation for evolve with prime numbers, 1 i32 input, 1 i32 output
-func perByteEvaluation_Primes(ind *cxast.CXProgram, solPrototype *cxast.CXFunction, numberOfRounds int) int64 {
+func perByteEvaluation_Primes(ind *cxast.CXProgram, solPrototype *cxast.CXFunction, numberOfRounds int) float64 {
 	var points int64 = 0
 	var tmp *cxast.CXProgram = cxast.PROGRAM
 	cxast.PROGRAM = ind
@@ -55,5 +55,5 @@ func perByteEvaluation_Primes(ind *cxast.CXProgram, solPrototype *cxast.CXFuncti
 	}
 
 	cxast.PROGRAM = tmp
-	return points
+	return float64(points)
 }
